Reorganized components folder
