+++ resources/assets/js/components/layout/app-header.vue
